% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.conditional.R
\name{estimate.conditional}
\alias{estimate.conditional}
\title{Computes moped estimated conditional distribution function values for X|Y1,Y2,...}
\usage{
estimate.conditional(
  fit,
  X = NULL,
  K.X = NULL,
  Y,
  K.Y = NULL,
  X.variable = NULL,
  Y.variables = NULL
)
}
\arguments{
\item{fit}{\code{moped} type variable outputted from \code{moped()}.}

\item{X}{Vector of values with which to estimate conditional probabilities.
\code{X} must have length consistent with number of observations in \code{Y}.
\code{X} is optional when reference density of X variable is "Uniform" and only
approximation polynomial coefficients are to be determined.}

\item{K.X}{Integer maximum polynomial order of approximation on X variable.
Must be less than or equal to the maximum MPO K specified in \code{moped()}.
The default is the \code{opt_mpo} or \code{KMax} (if \code{opt_mpo = NULL}) specified
in \code{fit}.}

\item{Y}{A data frame in which to look for conditional (Y) variables with which
to estimate probability values. Must contain column names matching variables
in the \code{moped} object.}

\item{K.Y}{Integer vector maximum polynomial order of approximation on each
conditional variable. Must be less than or equal to the maximum MPO K
specified in \code{moped()}. The default is the \code{opt_mpo} or \code{KMax}
(if \code{opt_mpo = NULL}) specified in \code{fit}.}

\item{X.variable}{Integer or character string of variable name corresponding
to the \code{moped} position or column name of the variable to be predicted from
\code{moped} object. The default is 1 (first variable in \code{fit}).}

\item{Y.variables}{Integer vector or character string of variable names
corresponding to the \code{moped} position or column name of the variable(s) to
be conditioned on from \code{moped} object. If \code{NULL} conditions on all non
\code{X.variable} variables.}
}
\value{
\code{estimate.conditional()} returns a list with the following components:
\itemize{
\item \code{Prob} - vector of computed probabilities when X is specified.
\item \code{coef} - An array of coefficients of the polynomial approximation.
When variable reference density is "Uniform", coefficients
are not specific for each value of X.
}
}
\description{
\code{estimate.conditional()} is used to compute moped estimated conditional
distribution function values for a single variable X given a data frame of
conditional Y values.
}
\examples{
require(ISLR)
Data_full <- Wage

require(tidyverse)
Data <- Data_full \%>\%
select(age, education, jobclass,wage)

# Convert Categorical Data to Continuous Data
Data_x <- make.cont(Data, catvar = 2:3)

# Fitting multivariate orthogonal polynomial based
# density estimation function

# Fitting the Data
Fit <- moped(Data_x)

# Compute moped conditional distribution estimate
cond_pred <- estimate.conditional(Fit,
X=seq(20,300,length.out=100),
Y = x0[rep(1,100),-4],
K.Y=rep(7,3),
K.X=7,
X.variable = "wage",
Y.variables = c("age","education","jobclass"))
}
