% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.marg.cdf.R
\name{estimate.marg.cdf}
\alias{estimate.marg.cdf}
\title{Calculate univariate marginal distribution function probabilities.}
\usage{
estimate.marg.cdf(fit, X = NULL, K = NULL, nprobs = 1, variable = 1)
}
\arguments{
\item{fit}{moped type variable outputted from \code{moped()}.}

\item{X}{vector of values to predict the marginal probability. Not required
for variables with "Uniform" reference densities to compute coefficients.}

\item{K}{integer maximum polynomial order of approximation on marginal
variable. Must be less than or equal to the maximum MPO K specified in
\code{moped()}.The default is the \code{opt_mpo} or \code{KMax} (if \code{opt_mpo = NULL}) specified
in \code{fit}.}

\item{nprobs}{integer of number of probability coefficients replications to be
outputted. Used when \code{X = NULL} for variables with \code{"Uniform"} reference densities.
The default value is 1 (no replications).}

\item{variable}{integer or string of variable name corresponding to which marginal
variable position or name to be predicted from \code{moped} object.}
}
\value{
\code{estimate.marg.cdf()} returns a list with the following components:
\itemize{
\item \code{Prob} - vector of computed probabilities when X is specified.
\item \code{coef} - An array of coefficients of the polynomial approximation.
When variable reference density is "Uniform", coefficients
are not specific for each value of X.
}
}
\description{
\code{estimate.marg.cdf()} is used to calculate univariate marginal cumulative
distribution function probabilities and the polynomial coefficients of the
polynomical approximation from any \code{moped} object.
}
\examples{
require(ISLR)
Data_full <- Wage

require(tidyverse)
Data <- Data_full \%>\%
select(age, education, jobclass,wage)

# Convert Categorical Data to Continuous Data
Data_x <- make.cont(Data, catvar = 2:3)

# Fitting multivariate orthogonal polynomial based
# density estimation function

# Fitting the Data
Fit <- moped(Data_x)

# Compute marginal distribution function probabilities of "wage"
x <- seq(21,310,length.out = 100)
wage_prob <- estimate.marg.cdf(Fit, X = x, K = 10, variable = "wage")
}
