% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cat.R
\name{make.cat}
\alias{make.cat}
\title{Convert previously continuous-transformed variables back to categorical variables}
\usage{
make.cat(Sample, fit = NULL)
}
\arguments{
\item{Sample}{A data frame outputted from \code{make.cont()} or \code{m.resample()} which
originally contained categorical variables.}

\item{fit}{optional \code{moped} type object outputted from \code{moped()}. Must be
specified if \code{Sample} has lost \code{Cats} attribute due to subsetting or
transformation.}
}
\value{
\code{make.cat()} returns a data frame.
}
\description{
Convert previously continuous-transformed variables back to categorical variables
}
\examples{
require(ISLR)
Data_full <- Wage

require(tidyverse)
Data <- Data_full \%>\%
select(age, maritl, race, education, jobclass, wage)
Data_amal <- make.cont(Data,catvar = c("maritl", "race", "education", "jobclass"),
amalgams = list(1:2,3:4))

# Convert previously continuous-transformed variables back to categorical variables.
make.cat(Data_amal)

# Categorised resampled data from `moped()`.
Data <- Data_full \%>\%
  select(education, jobclass, wage)
Data_numeric <- make.cont(Data, catvar = c("education","jobclass"))
Fit <- moped(Data_numeric)
new_Data_num <- m.resample(Fit)
new_Data <- make.cat(new_Data_num)

# If data has been subsetted. Fit is also required.
new_subset_Data <- make.cat(new_Data_num[1:10,-3],fit = Fit)
}
