% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cont.R
\name{make.cont}
\alias{make.cont}
\title{Convert categorical variables into continuous.}
\usage{
make.cont(Sample, catvar = 1:NCOL(Sample), amalgams = NULL)
}
\arguments{
\item{Sample}{A data frame.}

\item{catvar}{Columns to convert into continuous variables. Variable names
can be used as if they were positions in the data frame.}

\item{amalgams}{A list using numeric positions indicating the order of variables in
\code{"catvar"}. Each list item contains a vector of positions of \code{"catvar"}
that are to be amalgamated into a single continuous variable.}
}
\value{
\code{make.cont()} returns a data frame with strictly continuous values.
}
\description{
\code{make.cont()} is used to converting categorical variables into continuous
variables. It must done before fitting data to the density estimation
function \code{moped()}.
}
\examples{
library(tidyverse)
Data_full <- ISLR::Wage

# Convert Categorical Data to Continuous Data
Data <- Data_full \%>\%
select(age, education, jobclass, wage)
Data_x <- make.cont(Data, catvar = 2:3)
# Select variables by name
Data_x <- make.cont(Data, catvar = c("education", "jobclass"))

# Convert categorical with amalgamations
Data <- Data_full \%>\%
select(age, maritl, race, education, jobclass, wage)
Data_amal <- make.cont(Data,catvar = c("maritl", "race", "education", "jobclass"),
amalgams = list(
1:2, #maritl and race are amalgamated into a single variable.
3:4 #education and jobclass are amalgamated into a single variable.
))
}
