% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.plot.R
\name{marginal.plot}
\alias{marginal.plot}
\title{Generate marginal plots for each variable with a range of polynomial orders K}
\usage{
marginal.plot(fit, k.range = 1:fit$KMax, ncol = 4, prompt = TRUE)
}
\arguments{
\item{fit}{A \code{moped} object from \code{moped()}.}

\item{k.range}{Range of polynomial orders K. Default is 1 to \code{KMax} specified
in \code{moped} object.}

\item{ncol}{Integer vector of number of columns the marginal plots shown for
a given variable.}

\item{prompt}{Logical, whether it needs the prompt to show the next plot.
If \code{TRUE} (the default), press enter to show the next plot, If \code{FALSE},
outputs all plots simultaneously.}
}
\value{
\code{marginal.plot()} returns plots of each marginal variable.
}
\description{
\code{marginal.plot()} produces marginal plots for each variable with a range of
polynomial orders K. Optimal bin width uses the Freedman-Diaconis rule for
each variable.
}
\examples{
require(sdcMicro)
Data <- CASCrefmicrodata[,c(2,3,4,6)]

# Fitting multivariate orthogonal polynomial based
# density estimation function using default setting
Fit <- moped(Data)

# Check marginal densities with different polynomial K
# Default setting, showing all marginal densities with prompting
marginal.plot(Fit)

# Marginal densities with user-specified polynomial order, number of columns
# grid per plot and no prompting required
marginal.plot(Fit, k.range = 3:8, ncol =3, prompt = FALSE)
}
