% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{polynomial}
\alias{polynomial}
\title{Computing the univariate hyper-geometric type orthogonal polynomials.}
\usage{
polynomial(fit, X, K = NULL, variables = NULL)
}
\arguments{
\item{fit}{\code{moped} type object outputted from \code{moped()}.}

\item{X}{data frame of observations with column names matching selected
variables from \code{moped} type object.}

\item{K}{Integer vector of max polynomial order of approximation on each
variable. Must be less than or equal to the maximum MPO K specified in
\code{moped()}. The default is the \code{opt_mpo} or \code{KMax} (if \code{opt_mpo = NULL}) specified
in \code{fit}.}

\item{variables}{Integer vector or character string of variable names. The
\code{moped} position or column name of the variable(s) to be predicted from
\code{moped} object. The default is \code{1:fit$Nv}.}
}
\value{
A list with the following components:
\itemize{
\item \code{P} - List of marginal hyper-geometric orthogonal polynomial values.
\item \code{PdfTerms} - Vector of reference density values for each obs of X.
}
}
\description{
\code{polynomial()} is used to calculate the univariate hyper-geometric
type orthogonal polynomials used in the \code{moped()} function.
}
\examples{
Data_full <- ISLR::Wage
Data <- Data_full \%>\%
select(age, education, jobclass,wage)

# Convert Categorical Data to Continuous Data
Data_x <- make.cont(Data, catvar = 2:3)

# Fitting multivariate orthogonal polynomial based
# density estimation function
# Requires a data frame of bounds to fit on data.
bounds <- data.frame(
age  = c(18,80),
education = c(0,1),
jobclass = c(0,1),
wage = c(0,350)
)

# Fitting the Data
Fit <- moped(
Data_x,
K=10,
Distrib = rep("Uniform", 4),
bounds = bounds,
variance = T,
recurrence = F,
parallel = F,
ncores = NULL,
mpo = T
)

# Define the observation which the probability is desired.
x0 <- Data_x[1:2,]

# Manually Compute MBD Polynomials
polynomial(Fit, X=x0, K=7)
}
