% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.resample.R
\name{m.resample}
\alias{m.resample}
\title{Generate synthetic samples}
\usage{
m.resample(
  fit,
  K = NULL,
  variables = NULL,
  Sample = fit$SampleStats$Sample,
  n = NROW(Sample),
  bounds = NULL,
  replicates = 1,
  parallel = F,
  ncores = NULL,
  mps = 5000,
  fixed.var = NULL,
  er_alert = T
)
}
\arguments{
\item{fit}{MBDensity Type Variable. Outputed from MBDensity.}

\item{K}{Integer vector. Maximum Truncation of Approximation on each variable.
The default is the maximum MPO Order in \code{moped} object.}

\item{variables}{Integer vector or character string. Variables to be
predicted from \code{moped} object. The default is 1:Nv or 1:NCOL(Sample)
whichever smallest.}

\item{Sample}{A data frame defines the size of the synthetic data frame.}

\item{n}{Integer vector. The number of rows in the sample data frame.}

\item{bounds}{A data frame. Bounds allows you to control the grid min and max.
Should be an array of 2 x number of variables. \code{NULL} is the default.}

\item{replicates}{Integer vector. The default is 1.}

\item{parallel}{Logical. If \code{FALSE} (the default), parallel computing is not
used.}

\item{ncores}{Integer vector. NCores to use in parallel computing.}

\item{mps}{Integer vector. The default is 5000. (?Brad)}

\item{fixed.var}{The position of the variable(s) conditional upon. The
default is \code{NULL}.}

\item{er_alert}{Logical. The default is \code{TRUE}. (?Brad)}
}
\value{
\code{m.resample()} returns a data frame.
}
\description{
\code{m.resample()} is used to genderate synthetic samples.
}
\examples{
Data_full <- ISLR::Wage
Data <- Data_full \%>\%
select(age, education, jobclass,wage)

# Convert Categorical Data to Continuous Data
Data_x <- make.cont(Data, catvar = 2:3)

# Fitting multivariate orthogonal polynomial based
# density estimation function
# Requires a data frame of bounds to fit on data.
bounds <- data.frame(
age  = c(18,80),
education = c(0,1),
jobclass = c(0,1),
wage = c(0,350)
)

# Fitting the Data
Fit <- moped(
Data_x,
K=10,
Distrib = rep("Uniform", 7),
bounds = bounds,
variance = T,
recurrence = F,
parallel = F,
ncores = NULL,
mpo = T
)

# Generating resampled Obs
resampled <- m.resample(Fit, K=3, Sample=Data_x, fixed.var = 1, replicates = 1)
resampled <- m.resample(Fit, K=3, replicates = 1) # A fully resampled dataset
resampled <- m.resample(Fit, K=3, Sample = Data_x, variables = 4, replicates = 1)

# Marginal Synthetic
resampled_marginal <- m.resample(Fit,
Sample = Data_x[,3:4],
K = c(3,4),
variables = 3:4,
replicates = 1
)

# Convert previously continulized variables back to categorical variables.
resampled <- make.cat(resampled)




}
