% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.resample.R
\name{m.resample}
\alias{m.resample}
\title{Generate resampled samples from 'moped' estimate.}
\usage{
m.resample(
  fit,
  K = NULL,
  variables = 1:fit$Nv,
  Sample = fit$SampleStats$Sample,
  n = NROW(Sample),
  bounds = NULL,
  replicates = 1,
  parallel = F,
  ncores = NULL,
  mps = 5000,
  fixed.var = NULL,
  er_alert = T
)
}
\arguments{
\item{fit}{moped type variable outputted from \code{moped()}.}

\item{K}{Integer vector of max polynomial order of approximation on each
variable. Must be less than or equal to the maximum MPO K specified in
\code{moped()}. The default is the \code{opt_mpo} or \code{KMax} (if \code{opt_mpo = NULL})
specified in \code{fit}.}

\item{variables}{Integer vector or character string of variable names. The
\code{moped} position or column name of the variable(s) to be predicted from
\code{moped} object. The default is \code{1:fit$Nv}.}

\item{Sample}{A data frame of initial values used to impute values. Must
contain column names matching variables in the \code{moped} object. Default is
the Sample used to fit the \code{moped} object.}

\item{n}{Integer vector of the number of rows to be simulated.}

\item{bounds}{An optional data frame specifying the limits to be used on
bounded space. Should be an array of 2 x number of variables with each
column having the lower and upper limit.}

\item{replicates}{Integer vector determining the number of complete Gibbs
sampling passes to be performed.}

\item{parallel}{Logical that if \code{TRUE} uses the \code{parallel} package to simulate
values using parallel computing.}

\item{ncores}{Integer vector that determines the number of cores used in
parallel computing.}

\item{mps}{Integer vector that places a limit on maximum number of probabilities
calculated at a time. The default is 5000.}

\item{fixed.var}{Integer vector or string of variable names detailing the
\code{moped} position or column name of the variable(s) conditioned upon without
imputation. The default is \code{NULL}.}

\item{er_alert}{Logical that if \code{TRUE} (default) will return an error
message when observations require re-sampling due to errors.}
}
\value{
\code{m.resample()} returns a data frame of imputed values.
}
\description{
\code{m.resample()} is used to generate synthetic samples from full, conditional,
or marginal moped density estimate.
}
\examples{
require(sdcMicro)
Data <- CASCrefmicrodata[,c(2,3,4,6)]

Fit <- moped(Data)

# Generating resampled (synthetic) observations
# Sample 100 obs from moped joint density estimate without updating "age"
m.resample(Fit, K = 3, Sample = Data_x, n = 100, fixed.var = "age")

# Simulate a fully resampled data set of same size as Data_x.
m.resample(Fit, K = 3, replicates = 2) # 2 Gibbs passes used.

# Simulate a fully resampled data set of same size as Data.
m.resample(Fit, K=3, replicates = 2) # 2 Gibbs passes used.

# Sample 100 obs from moped joint density estimate without updating "AGI"
m.resample(Fit, K=3, Sample=Data, n = 100, fixed.var = "AGI")

# Sample values for "AGI" conditional on other three variables from Data
m.resample(Fit, variables =1)

# Sample fully synthetic data set from marginal bivariate moped density
# estimate of "AGI" and "STATETAX"
m.resample(Fit, Sample = Data[,c(1,4)], K = c(4,5), variables = c(1,4),
           replicates = 1)
}
