% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cat.R
\name{make.cat}
\alias{make.cat}
\title{Convert previously continulized variables back to categorical variables.}
\usage{
make.cat(Sample)
}
\arguments{
\item{Sample}{A data frame.}
}
\value{
\code{make.cat()} returns a data frame.
}
\description{
Convert previously continulized variables back to categorical variables.
}
\examples{
Data_full <- ISLR::Wage

Data <- Data_full \%>\%
select(age, maritl, race, education, jobclass, wage)
Data_amal <- make.cont(Data,catvar = c("maritl", "race", "education", "jobclass"),
amalgams = list(1:2,3:4))

# Convert previously continulized variables back to categorical variables.
make.cat(Data_amal)
}
