% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cont.R
\name{make.cont}
\alias{make.cont}
\title{Convert categorical variables into continuous.}
\usage{
make.cont(Sample, catvar = 1:NCOL(Sample), amalgams = NULL)
}
\arguments{
\item{Sample}{A data frame.}

\item{catvar}{Columns to convert into continuous variables. Variable names
can be used as if they were positions in the data frame.}

\item{amalgams}{A list using numeric indicating the order of variables in
\code{"catvar"}. It is the structure of amalgamating categorical variable when
being continulized.}
}
\value{
\code{make.cont()} returns a data frame.
}
\description{
\code{make.cont()} is used to converting categorical variables into continuous
variables. It must done before fitting data to the density estimation
function \code{moped()}.
}
\examples{
Data_full <- ISLR::Wage

# Convert Categorical Data to Continuous Data
Data <- Data_full \%>\%
select(age, education, jobclass, wage)
Data_x <- make.cont(Data,catvar = 2:3)
# Select variables by name
Data_x <- make.cont(Data,catvar = c("education","jobclass"))

# Convert Categorical with amalgamations
Data <- Data_full \%>\%
select(age, maritl, race, education, jobclass, wage)
Data_amal <- make.cont(Data,catvar = c("maritl","race","education","jobclass"),
amalgams = list(1:2,3:4))

}
