% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.plot.R
\name{marginal.plot}
\alias{marginal.plot}
\title{Marginal plots for each variable with a range of polynomial orders K.}
\usage{
marginal.plot(fit, k.range = 1:min(fit$KMax), ncol = 4, prompt = TRUE)
}
\arguments{
\item{fit}{A result object from \code{moped()}.}

\item{k.range}{Range of polynomial orders K.}

\item{ncol}{Integer vector. Number of columns the marginal plots shown for a given variable.}

\item{prompt}{Logical, whether it needs the prompt to show the next plot.
If \code{TRUE} (the default), press enter to show the next plot, If \code{FALSE},
it means showing all plots simultaneously.}
}
\description{
\code{marginal.plot()} is used to Marginal plots for each variable with a range of
polynomial orders K. Optimal bin width using the Freedman-Diaconis rule for
each variable.
}
\examples{
Data_full <- ISLR::Wage
Data <- Data_full \%>\%
select(age, education, jobclass,wage)

# Convert Categorical Data to Continuous Data
Data_x <- make.cont(Data, catvar = 2:3)

# Fitting multivariate orthogonal polynomial based
# density estimation function
# Requires a data frame of bounds to fit on data.
bounds <- data.frame(
age  = c(18,80),
education = c(0,1),
jobclass = c(0,1),
wage = c(0,350)
)

Fitting the Data
Fit <- moped(
Data_x,
K=10,
Distrib = rep("Uniform", 7),
bounds = bounds,
variance = T,
recurrence = F,
parallel = F,
ncores = NULL,
mpo = T
)

# Check marginal densities with different polynomial K
marginal.plot(Fit)
marginal.plot(Fit, k.range = 3:8, ncol =3, prompt = FALSE)
}
