% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.mpo.R
\name{validate.mpo}
\alias{validate.mpo}
\title{Estimate optimal max polynomial order with repeated k-fold cross-validation}
\usage{
\method{validate}{mpo}(fit, K = fit$KMax, nfolds = 5, repeats = 10, variance = T)
}
\arguments{
\item{fit}{\code{moped} type variable outputted from \code{moped()}.}

\item{K}{Integer of the maximum possible max polynomial order of approximation.
Default is \code{KMax} specified in \code{fit}.}

\item{nfolds}{Integer that determines the number of folds (k) to perform in
k-fold cross-validation.}

\item{repeats}{Integer that determines the number of times k-fold cross-validation
is repeated.}

\item{variance}{Logical that if \code{TRUE} (default), computes a variance estimate
of each coefficient.}
}
\value{
\code{validate.mpo()} returns a list containing:
\itemize{
\item \code{Cn} - Array of estimated moment-based coefficients.
\item \code{varCn} - Array of variance estimates for Cn. Computed if \code{variance = TRUE}.
\item \code{Nk_norm} - Array of estimated shifted Nk Norm values.
\item \code{opt_mpo_vec} - Estimated optimal max polynomial order where K is vector.
\item \code{opt_mpo} - Estimated optimal max polynomial order where K is constant.
}
}
\description{
\code{validate.mpo()} performs repeated k-fold cross-validation to
determine an unbiased estimate of the shifted \code{Nk}-Norm that when minimised, it
provides an estimate of the optimal max polynomial order. The coefficient
estimates are also calculated as well as  their corresponding variance
estimates (optional).
}
\examples{
require(sdcMicro)
Data <- CASCrefmicrodata[,c(2,3,4,6)]

# Fitting multivariate orthogonal polynomial based
# density estimation function using default setting
Fit <- moped(Data)

# Select the optimal polynomial order K
val <- validate.mpo(fit = Fit) #warning! it might take a while
val$opt_mpo_vec #show the vector of optimal polynomial orders
}
