% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.mpo.R
\name{validate.mpo}
\alias{validate.mpo}
\title{Perform repeated k-fold cross-validation to estimate optimal max polynomial order.}
\usage{
validate.mpo(fit, K = fit$KMax, nfolds = 5, repeats = 10, variance = T)
}
\arguments{
\item{fit}{\code{moped} type variable. Outputted from \code{moped()}.}

\item{K}{Integer. Maximum possible Max Polynomial Order of approximation.
Default is \code{KMax} specified in \code{moped} object.}

\item{nfolds}{Integer. If \code{opt.mpo = TRUE} number of folds (k) to perform in
k-fold cross-validation. Default is 5.}

\item{repeats}{Integer. If \code{opt.mpo = TRUE} number of times k-fold
cross-validation is repeated. Default is 10.}

\item{variance}{Logical. If \code{TRUE} (the default), a variance estimate of each
coefficient is calculated.}
}
\value{
\code{moped()} returns a list containing:
\itemize{
\item \code{Cn} - Array of estimated moment-based coefficients.
\item \code{varCn} - Array of variance estimates for Cn. Computed if \code{variance = TRUE}.
\item \code{Nk_norm} - Array of estimated shifted Nk Norm values.
Computed if \code{opt.mpo = TRUE}.
\item \code{opt_mpo_vec} - Estimated optimal max polynomial order where K is vector.
\item \code{opt_mpo} - Estimated optimal max polynomial order where K is constant.
}
}
\description{
\code{validate.mpo()} performs repeated k-fold cross-validation to determine
an unbiased estimate of the shifted Nk-Norm that when minimised provides an
estimate of the optimal max polynomial order. The coefficient estimates
are also calculated as is (optional) their corresponding variance estimates.
}
